feat(react-native): sample app & voice transcription (#7)
This PR looks larger than it actually is because of the introduction of
the RN sample app in the repo. Realistically, it's whatever the app's
`App.tsx` file contains and then everything within
`packages/react-native-sdk`.

Everything else does not need to be reviewed and was generated with
`create-react-native-app` (with slight modifications for known issues as
well as library installations).

In addition, it also includes:

- Clipboard support for code blocks
- A rudimentary state layer for bottom sheet, dictation and the text
input
- Refactoring of the services API a bit so that it's easier to consume

TODO: 

- [x] iOS voice dictation
- [x] Clean `useDictation` up a bit