feat(react-native): expo config plugin for dictation (#19)
This PR adds an `Expo` config to be built together with the SDK and
ready to be used for the purposes of dictation permissions. It will add
both the `Android` permissions to `AndroidManifest.xml` as well as the
`iOS` permissions to `Info.plist`.

In addition to that, it also adds an `Expo` sample app to be used for
testing purposes.